Finished the combat mechanics.
Added some UI that shows the Combatants Current and Max Health.
The virtual camera now stays focused on the middle of the room when combat starts. It goes back to following the player when combat ends.
The Player Movers and Room Spawners now don't appear when the player walks into an Enemy Room and combat starts.
The Playing Card Deck is reshuffled at the start of every round.